+++ spirit/core/static/spirit/scripts/bookmark.js
@@ -63,18 +63,22 @@
     };
 
     Bookmark.prototype.sendCommentNumber = function() {
-      var post;
+      var post, sentCommentNumber;
       if (this.mark.isSending) {
         return;
       }
       this.mark.isSending = true;
+      sentCommentNumber = this.mark.commentNumber;
       post = $.post(this.options.target, {
         csrfmiddlewaretoken: this.options.csrfToken,
         comment_number: this.mark.commentNumber
       });
       return post.always((function(_this) {
         return function() {
-          return _this.mark.isSending = false;
+          _this.mark.isSending = false;
+          if (_this.mark.commentNumber > sentCommentNumber) {
+            return _this.sendCommentNumber();
+          }
         };
       })(this));
     };